pages adjusted react env adjusted and prettier used
