+++ public/media-hunt-react-app/.env.example
@@ -1,0 +1,1 @@
+REACT_APP_TOKEN=
