--- public/media-hunt-react-app/src/.env.example
@@ -1,1 +1,0 @@
-TMDB_API_TOKEN=
