+++ public/media-hunt-react-app/src/App.jsx
@@ -20,6 +20,6 @@ function App() {
         <Routes />
       </main>
     </>
-  )
+  );
 }
 export default App;
