+++ public/media-hunt-react-app/src/Routes.js
@@ -11,10 +11,10 @@ export default function Routes() {
       <Route path="/" element={<Home />} />
       <Route path="Movie">
         <Route index element={<Movie />} />
-        <Route path="/Movie/movieDetails" element={<MovieDetails/>}/>
+        <Route path="/Movie/movieDetails" element={<MovieDetails />} />
       </Route>
       <Route path="Tv" index element={<Tv />} />
-      <Route path="/Tv/tvDetails" index element={<TvDetails/>}/>
+      <Route path="/Tv/tvDetails" index element={<TvDetails />} />
     </BaseRoutes>
   );
 }
