+++ public/media-hunt-react-app/src/index.css
@@ -1,4 +1,5 @@
-* {margin: 0;
+* {
+  margin: 0;
   padding: 0;
   box-sizing: border-box;
   font-family: 'Sora', sans-serif;
@@ -32,7 +33,7 @@ header nav ul li a {
   text-decoration: none;
 }
 
-header nav ul li + li {
+header nav ul li+li {
   margin-left: 24px;
 }
