--- public/media-hunt-react-app/src/pages/MovieDetails.js
@@ -1,6 +1,0 @@
-export default function MovieDetails () {
-
-    return (
-        <div>OVO je neka movie stranica</div>
-    )
-}
