+++ public/media-hunt-react-app/src/pages/MovieDetails.jsx
@@ -1,0 +1,3 @@
+export default function MovieDetails() {
+  return <div>OVO je neka movie stranica</div>;
+}
