--- public/media-hunt-react-app/src/pages/TvDetails.js
@@ -1,5 +1,0 @@
-export default function TvDetails () {
-    return (
-        <div>OVO je neka tv stranica</div>
-    )
-}
