+++ public/media-hunt-react-app/src/pages/TvDetails.jsx
@@ -1,0 +1,3 @@
+export default function TvDetails() {
+  return <div>OVO je neka tv stranica</div>;
+}
